+++ src/main/java/org/opensearch/searchrelevance/common/MetricsConstants.java
@@ -20,9 +20,23 @@ public class MetricsConstants {
     /**
      * pariwise comparison field names
      */
-    public static final String METRICS_PAIRWISE_COMPARISON_FIELD_NAME = "pairwiseComparison";
+    public static final String METRICS_PAIRWISE_COMPARISON_FIELD_NAME = "metrics";
     public static final String PAIRWISE_FIELD_NAME_A = "0";
     public static final String PAIRWISE_FIELD_NAME_B = "1";
+    public static final String PAIRWISE_FIELD_NAME_SNAPSHOTS = "snapshots";
+    public static final String PAIRWISE_FIELD_NAME_SEARCH_CONFIGURATION_ID = "searchConfigurationId";
+    public static final String PAIRWISE_FIELD_NAME_DOC_IDS = "docIds";
+    public static final String PAIRWISE_FIELD_NAME_METRIC = "metric";
+    public static final String PAIRWISE_FIELD_NAME_VALUE = "value";
+    public static final String PAIRWISE_FIELD_NAME_QUERY_TEXT = "queryText";
+
+    /**
+     * pointwise experiment field names
+     */
+    public static final String POINTWISE_FIELD_NAME_EVALUATION_ID = "evaluationId";
+    public static final String POINTWISE_FIELD_NAME_EXPERIMENT_VARIANT_ID = "experimentVariantId";
+    public static final String POINTWISE_FIELD_NAME_EVALUATION_RESULTS = "evaluationResults";
+    public static final String POINTWISE_FIELD_NAME_SEARCH_CONFIGURATION_ID = "searchConfigurationId";
 
     /**
      * metadata map fields