+++ src/main/java/org/opensearch/searchrelevance/metrics/EvaluationMetrics.java
@@ -7,6 +7,8 @@
  */
 package org.opensearch.searchrelevance.metrics;
 
+import static org.opensearch.searchrelevance.common.MetricsConstants.PAIRWISE_FIELD_NAME_METRIC;
+import static org.opensearch.searchrelevance.common.MetricsConstants.PAIRWISE_FIELD_NAME_VALUE;
 import static org.opensearch.searchrelevance.metrics.calculator.Evaluation.METRICS_MEAN_AVERAGE_PRECISION_AT;
 import static org.opensearch.searchrelevance.metrics.calculator.Evaluation.METRICS_NORMALIZED_DISCOUNTED_CUMULATIVE_GAIN_AT;
 import static org.opensearch.searchrelevance.metrics.calculator.Evaluation.METRICS_PRECISION_AT;
@@ -14,6 +16,7 @@ import static org.opensearch.searchrelevance.metrics.calculator.Evaluation.calcu
 import static org.opensearch.searchrelevance.metrics.calculator.Evaluation.calculateNDCGAtK;
 import static org.opensearch.searchrelevance.metrics.calculator.Evaluation.calculatePrecisionAtK;
 
+import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Locale;
@@ -24,12 +27,18 @@ import java.util.Map;
  */
 public class EvaluationMetrics {
 
+    private static void addMetric(List<Map<String, Object>> metrics, String metricName, double value) {
+        Map<String, Object> metric = new HashMap<>();
+        metric.put(PAIRWISE_FIELD_NAME_METRIC, metricName);
+        metric.put(PAIRWISE_FIELD_NAME_VALUE, value);
+        metrics.add(metric);
+    }
+
     /**
      * calculate evaluation metrics with evaluation calculators.
      */
-    public static Map<String, String> calculateEvaluationMetrics(List<String> docIds, Map<String, String> judgments, int k) {
-        Map<String, String> currSearchConfigMetrics = new HashMap<>();
-
+    public static List<Map<String, Object>> calculateEvaluationMetrics(List<String> docIds, Map<String, String> judgments, int k) {
+        List<Map<String, Object>> metrics = new ArrayList<>();
         List<String> docsWithScores = docIds.stream().filter(judgments::containsKey).toList();
 
         // calculate coverage statistics
@@ -38,15 +47,13 @@ public class EvaluationMetrics {
 
         double coverage = totalCount > 0 ? Math.round((double) totalDocsWithScores / totalCount * 100.0) / 100.0 : 0.0;
 
-        // docIds that don't have a score will be assumed irrelevant with a score of 0
-        // Coverage is the metric of what percentage of docIds per query have score
-        currSearchConfigMetrics.put(String.format(Locale.ROOT, "Coverage@%d", k), String.valueOf(coverage));
-        currSearchConfigMetrics.put(METRICS_PRECISION_AT + k, String.valueOf(calculatePrecisionAtK(docIds, judgments, k)));
-        currSearchConfigMetrics.put(METRICS_MEAN_AVERAGE_PRECISION_AT + k, String.valueOf(calculateMAPAtK(docIds, judgments, k)));
-        currSearchConfigMetrics.put(
-            METRICS_NORMALIZED_DISCOUNTED_CUMULATIVE_GAIN_AT + k,
-            String.valueOf(calculateNDCGAtK(docIds, judgments, k))
-        );
-        return currSearchConfigMetrics;
+        // TODO: it's not guarantee that each docId will have its score, especially for UBI data.
+        // Need to define a reliable rate. say, coverage > 80%, then the results become reliable
+        addMetric(metrics, String.format(Locale.ROOT, "Coverage@%d", k), coverage);
+        addMetric(metrics, METRICS_PRECISION_AT + k, calculatePrecisionAtK(docIds, judgments, k));
+        addMetric(metrics, METRICS_MEAN_AVERAGE_PRECISION_AT + k, calculateMAPAtK(docIds, judgments, k));
+        addMetric(metrics, METRICS_NORMALIZED_DISCOUNTED_CUMULATIVE_GAIN_AT + k, calculateNDCGAtK(docIds, judgments, k));
+
+        return metrics;
     }
 }